+++ src/aux-server/aux-web/webpack.common.js
@@ -9,6 +9,7 @@ const WorkboxPlugin = require('workbox-webpack-plugin');
 const CircularDependencyPlugin = require('circular-dependency-plugin');
 const webpack = require('webpack');
 const { merge } = require('webpack-merge');
+const WebpackAssetsManifest = require('webpack-assets-manifest');
 
 const commitHash = childProcess
     .execSync('git rev-parse HEAD')
@@ -34,11 +35,19 @@ function playerConfig() {
                 '..',
                 'aux-vm-browser',
                 'html',
-                'IframeEntry.ts'
+                'IframeEntry.js'
             ),
             'service-worker': path.resolve(
                 __dirname,
                 './shared/service-worker.ts'
+            ),
+            worker: path.resolve(
+                __dirname,
+                '..',
+                '..',
+                'aux-vm-browser',
+                'vm',
+                'WorkerEntry.ts'
             ),
         },
         plugins: [
@@ -137,6 +146,7 @@ function playerConfig() {
                     },
                 ],
             }),
+            new WebpackAssetsManifest(),
         ],
     });
 }
@@ -164,6 +174,7 @@ function commonPlugins() {
             GIT_TAG: JSON.stringify(latestTag),
             PROXY_CORS_REQUESTS: process.env.PROXY_CORS_REQUESTS !== 'false',
         }),
+        new webpack.NormalModuleReplacementPlugin(/^esbuild$/, 'esbuild-wasm'),
     ];
 }
 
@@ -236,7 +247,7 @@ function baseConfig() {
                     use: 'exports-loader?vg=vg',
                 },
                 {
-                    test: /\.(gltf|glb)$/,
+                    test: /\.(gltf|glb|wasm)$/,
                     use: [
                         {
                             loader: 'file-loader',