+++ src/aux-vm/vm/test/TestAuxVM.ts
@@ -25,7 +25,7 @@ import union from 'lodash/union';
 import { AuxUser } from '../../AuxUser';
 import { StoredAux } from '../../StoredAux';
 import { ChannelActionResult } from '../../vm';
-import { BotStateUpdates } from '@casual-simulation/aux-common/aux-format-2';
+import { PortalEvent } from '../PortalEvents';
 
 export class TestAuxVM implements AuxVM {
     private _stateUpdated: Subject<StateUpdatedEvent>;
@@ -43,6 +43,7 @@ export class TestAuxVM implements AuxVM {
     connectionStateChanged: Subject<StatusUpdate>;
     versionUpdated: Subject<RuntimeStateVersion>;
     onError: Subject<AuxChannelErrorType>;
+    portalEvents: Subject<PortalEvent[]>;
     grant: string;
     user: AuxUser;
 
@@ -74,6 +75,15 @@ export class TestAuxVM implements AuxVM {
         this.connectionStateChanged = new Subject<StatusUpdate>();
         this.onError = new Subject<AuxChannelErrorType>();
         this.versionUpdated = new Subject<RuntimeStateVersion>();
+        this.portalEvents = new Subject();
+    }
+
+    registerCustomPortal(id: string): Promise<void> {
+        throw new Error('Method not implemented.');
+    }
+
+    updatePortalSource(id: string, source: string): Promise<void> {
+        throw new Error('Method not implemented.');
     }
 
     async shout(