+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CasualOS Changelog
 
+## V1.4.0
+
+#### Date: TBD
+
+### :rocket: Improvements
+
+-   Added an initial implementation of custom portals.
+    -   Custom portals are a way to write scripts that can interact directly with the web browser. This gives you the ability to do anything that is possible from inside a web browser.
+    -   The following functions are now available:
+        -   `portal.register(portalID, options?)`
+        -   `portal.addEntryPoint(portalID, tag)`
+        -   See the documentation for usage information.
+
 ## V1.3.14
 
 #### Date: 1/25/2021